Update the Document and Element interface to support a findFirst(...) set of methods similar to the old queryPath(...) but that only return the first matching element.
Renamed/deprecated the queryPath(...) methods to instead be find(...)
and updated across the driver.

Added a get(Class,String) method to the Document interface for fast,
type safe retrieval of document elements.

Updated the site to include Cobertura results and reordered the reports.

Updated the driver to exclude replica set members not in a primary or
secondary mode and to avoid members that fail to respond to a
replica-set status command.