Update the Server to track the maximum BSON object size allowed by the server and the version of the server.
Moved the BSON object Size to the socket connection and be based on the
server reported maximum size instead of the Client default.

Updated the Message interface and socket connection to be able to
quickly fail messages that require a specific version but the server's
version is before (less than) that version.