do not set 'myNumberToReturn' to negative number
There have been a number of significant changes in the product that were introduced in versions 3.0 and 3.2 that were released in 2015.  One such change is enhancements to our wire protocol.  Previously the legacy `OP_QUERY` spec had a single parameter, `ntoreturn`, which was overloaded to be used with both `limit` and `batchSize` options.  This behavior is described in our [legacy CRUD API specification](https://github.com/mongodb/specifications/blob/master/source/crud/crud.rst#combining-limit-and-batch-size-for-the-wire-protocol).  The ambiguity in terms of what the client was expecting when `ntoreturn` was provided resulted in suboptimal handling of the request in some situations.